Created header for event visitors
