--- descartes_light/solvers/include/descartes_light/solvers/bgl/event_visitor.h
@@ -1,4 +1,0 @@
-#ifndef EVENT_VISITOR_H
-#define EVENT_VISITOR_H
-
-#endif  // EVENT_VISITOR_H
